requirements.txt
flask==1.1.1
gunicorn
scikit-learn==0.24.2
numpy
pandas
pdfrw
reportlab
PyPDF2
pdfminer.six
pikepdf
flask_cors
celery[redis]==4.4.0
coverage==5.5
pytest
pytest-mock
requests-mock
pytest-celery
coverage2clover